added gcc check condition

+++ Peregrine/run_tests_ci.sh
@@ -5,9 +5,11 @@ for file in *
 do
     echo $file ...
     ../peregrine compile $file -emit-c
-    gcc temp.c -o output.o
-    chmod +x ./output.o
-    ./output.o
-    rm ./output.o
+    if gcc temp.c -o output.o then
+        echo "::error GCC FAILED"
+    else
+        chmod +x ./output.o
+        ./output.o
+        rm ./output.o
     
 done
